Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    float x,area;
-    scanf("%f",&x);
-    area = 3.14*x**2;
+    int x,area;
+    scanf("%d",&x);
+    area = 3.14*(x**2);
     printf("Area: %.2f\n", (float)area);
     return 0;
 }
